% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.R
\name{get_pixels}
\alias{get_pixels}
\title{Gets Pixels using a Shiny Gadget}
\usage{
get_pixels(grid = c(28, 28), size = c(500, 500), brush = matrix(c(0, 1, 1,
  1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0), 5, 5))
}
\arguments{
\item{grid}{The grid dimensions specified as a vector.}

\item{size}{The canvas dimensions specified as a vector.}

\item{brush}{The brush specified as a matrix.}
}
\description{
Creates an ShinyGadget to retrieve pixels as a vector.
}
