% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.R
\docType{package}
\name{pixels-package}
\alias{pixels}
\alias{pixels-package}
\title{Tools for Working with Image Pixels}
\description{
This package provides tools to show and draw image pixels using 'HTML' widgets 
and 'Shiny' applications.
}
\details{
This package can be used to: Visualize the 'MNIST' dataset for handwritten
digit recognition, create simple test cases for image classification, or
to create new datasets by building a 'Shiny' applicaiton that enables
many users to capture images for supervised learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/javierluraschi/pixels}
  \item Report bugs at \url{https://github.com/javierluraschi/pixels/issues}
}

}
\author{
\strong{Maintainer}: Javier Luraschi \email{javier@rstudio.com}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
