% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.R
\name{shiny_pixels_output}
\alias{shiny_pixels_output}
\title{Shiny Widget Output}
\usage{
shiny_pixels_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The identifier for this widget.}

\item{width}{The width for this widget.}

\item{height}{The height for this widget.}
}
\description{
Provides a Shiny Widget for Output.
}
