% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.R
\name{shiny_render_pixels}
\alias{shiny_render_pixels}
\title{Shiny Widget Render}
\usage{
shiny_render_pixels(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The \code{expr} for \code{shinyRenderWidget}.}

\item{env}{The \code{env} for \code{shinyRenderWidget}.}
}
\description{
Renders the Shiny Widget.
}
