% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels.R
\name{show_pixels}
\alias{show_pixels}
\title{Show Pixels Widget}
\usage{
show_pixels(pixels = NULL, grid = c(28, 28), size = c(500, 500),
  brush = matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
  0, 1, 1, 1, 0), 5, 5))
}
\arguments{
\item{pixels}{The pixels to render as a 1-dimensional vector, row-first
order expected.}

\item{grid}{The grid dimensions specified as a vector.}

\item{size}{The canvas dimensions specified as a vector.}

\item{brush}{The brush specified as a matrix.}
}
\description{
Creates an HTMLWidget to interact with pixels.
}
